refactor cli into command modules
